Simplify telemetry in ChatService
